refactor: Remove Association from name
BREAKING CHANGE
