+++ packages/cozy-client/src/associations/helpers.js
@@ -2,8 +2,8 @@ import pick from 'lodash/pick'
 import pickBy from 'lodash/pickBy'
 import Association from './Association'
 import HasOneInPlace from './HasOneInPlace'
-import HasMany from './HasManyAssociation'
-import HasManyFiles from './HasManyFilesAssociation'
+import HasMany from './HasMany'
+import HasManyFiles from './HasManyFiles'
 
 export const pickTypeAndId = x => pick(x, '_type', '_id')
 const applyHelper = (fn, objOrArr) =>
@@ -18,11 +18,12 @@ export const responseToRelationship = response =>
   })
 
 const aliases = {
-  'io.cozy.files:has-many': HasManyFilesAssociation,
-  'has-many': HasManyAssociation,
+  'io.cozy.files:has-many': HasManyFiles,
+  'has-many': HasMany,
   'belongs-to-in-place': HasOneInPlace,
   'has-one-in-place': HasOneInPlace
 }
+
 
 export const getClass = (doctype, type) => {
   if (typeof type !== 'string') {
