+++ packages/cozy-client/src/associations/index.js
@@ -1,5 +1,6 @@
-import HasManyFilesAssociation from './HasManyFilesAssociation'
-import HasManyAssociation from './HasManyAssociation'
+
+import HasManyFiles from './HasManyFiles'
+import HasMany from './HasMany'
 import HasOneInPlace from './HasOneInPlace'
 import Association from './Association'
 
@@ -7,7 +8,7 @@ export { responseToRelationship, getClass, create } from './helpers'
 
 export {
   Association,
-  HasManyFilesAssociation,
-  HasManyAssociation,
+  HasManyFiles,
+  HasMany,
   HasOneInPlace
 }
